Refactored HistoryScreen: Created HistoryScreenUI file and extracted UI for better preview support.
